+++ src/slog.rs
@@ -7,13 +7,13 @@ use std::{fmt, io};
 
 /// Key-Separator-Value serializer
 // Copied from `slog-stdlog`
-struct KSV<W: io::Write> {
+struct Ksv<W: io::Write> {
     io: W,
 }
 
-impl<W: io::Write> KSV<W> {
+impl<W: io::Write> Ksv<W> {
     fn new(io: W) -> Self {
-        KSV { io }
+        Ksv { io }
     }
 
     fn into_inner(self) -> W {
@@ -21,7 +21,7 @@ impl<W: io::Write> KSV<W> {
     }
 }
 
-impl<W: io::Write> slog::Serializer for KSV<W> {
+impl<W: io::Write> slog::Serializer for Ksv<W> {
     fn emit_arguments(&mut self, key: slog::Key, val: &fmt::Arguments) -> slog::Result {
         write!(self.io, ", {}: {}", key, val)?;
         Ok(())
@@ -48,7 +48,7 @@ impl<'a> fmt::Display for LazyLogString<'a> {
         write!(f, "{}", self.info.msg())?;
 
         let io = io::Cursor::new(Vec::new());
-        let mut ser = KSV::new(io);
+        let mut ser = Ksv::new(io);
 
         self.logger_values
             .serialize(self.info, &mut ser)