+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+next:
+- Update rust edition in Cargo.toml to 2021
+
 0.6.6:
 - Add functions to format output of log data as discussed in [issue #19](https://github.com/gin66/tui-logger/issues/19)
   The functions are with their default values: